non lib files for float to double move

+++ RayTrace/test_Matrix.cpp
@@ -20,16 +20,16 @@ using namespace ray;
 //    And M[3,2] = 15.5
 TEST(Matrix, Ctor4) {
 	Matrix4 m1;
-	std::array<float, 16> a1 = m1.to_array();
+	std::array<double, 16> a1 = m1.to_array();
 	EXPECT_EQ(a1[0], 1);
 
-	Matrix4 m2(std::array<float,16>{ 
+	Matrix4 m2(std::array<double,16>{ 
 		1,2,3,4,
 			5.5, 6.5, 7.5,  8.5,
 			9, 10, 11, 12,
 			13.5, 14.5, 15.5,16.5 
 	});
-	std::array<float, 16> a2 = m2.to_array();
+	std::array<double, 16> a2 = m2.to_array();
 	EXPECT_EQ(a2[0], 1);
 	EXPECT_EQ(a2[3], 4);
 	EXPECT_EQ(a2[4], 5.5);
@@ -50,7 +50,7 @@ TEST(Matrix, Ctor4) {
 TEST(Matrix, Ctor2) {
 	Matrix2 m1;
 	EXPECT_EQ(m1.m00, 1);
-	Matrix2 m2(std::array<float, 4>{
+	Matrix2 m2(std::array<double, 4>{
 		-3, 5, 1, -2
 	});
 	EXPECT_EQ(m2.m00, -3);
@@ -70,7 +70,7 @@ TEST(Matrix, Ctor2) {
 TEST(Matrix, Ctor3) {
 	Matrix3 m1;
 	EXPECT_EQ(m1.m00, 1);
-	Matrix3 m2(std::array<float, 9>{
+	Matrix3 m2(std::array<double, 9>{
 		-3, 5, 0,
 		1, -2, -7,
 		0,  1,  1
@@ -93,13 +93,13 @@ TEST(Matrix, Ctor3) {
 //      | 5 | 4 | 3 | 2 |
 //  Then A = B
 TEST(Matrix, Equal4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		1, 2, 3, 4,
 			5, 6, 7, 8,
 			9, 8, 7, 6,
 			5, 4, 3, 2
 	});
-	Matrix4 b(std::array<float, 16>{
+	Matrix4 b(std::array<double, 16>{
 		1, 2, 3, 4,
 			5, 6, 7, 8,
 			9, 8, 7, 6,
@@ -121,13 +121,13 @@ TEST(Matrix, Equal4) {
 //      | 4 | 3 | 2 | 1 |
 //  Then A != B
 TEST(Matrix, NotEqual4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		1, 2, 3, 4,
 			5, 6, 7, 8,
 			9, 8, 7, 6,
 			5, 4, 3, 2
 	});
-	Matrix4 b(std::array<float, 16>{
+	Matrix4 b(std::array<double, 16>{
 	 2, 3, 4, 5, 
 	 6, 7, 8, 9, 
 	 8, 7, 6, 5, 
@@ -153,13 +153,13 @@ TEST(Matrix, NotEqual4) {
 //      | 40|  58 | 110 | 102 |
 //      | 16|  26 |  46 |  42 |
 TEST(Matrix, MulMatrix4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		1, 2, 3, 4,
 			5, 6, 7, 8,
 			9, 8, 7, 6,
 			5, 4, 3, 2
 	});
-	Matrix4 b(std::array<float, 16>{
+	Matrix4 b(std::array<double, 16>{
 		-2, 1, 2, 3,
 		3, 2, 1, -1,
 		4, 3, 6, 5,
@@ -167,7 +167,7 @@ TEST(Matrix, MulMatrix4) {
 	});
 	Matrix4 result = a * b;
 
-	Matrix4 expect(std::array<float, 16>{
+	Matrix4 expect(std::array<double, 16>{
 		20, 22, 50, 48,
 		44, 54, 114,108,
 		40, 58, 110,102,
@@ -185,7 +185,7 @@ TEST(Matrix, MulMatrix4) {
 //    And b ← tuple(1, 2, 3, 1)
 //  Then A * b = tuple(18, 24, 33, 1)
 TEST(Matrix, MulVec3) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		1, 2, 3, 4,
 			2, 4, 4, 2,
 			8, 6, 4, 1,
@@ -204,7 +204,7 @@ TEST(Matrix, MulVec3) {
 //    | 4 | 8 | 16 | 32 |
 //  Then A * identity_matrix = A
 TEST(Matrix, MulIdent) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		0, 1, 2, 4,
 		1,2,4,8,
 		2,4,8,16,
@@ -237,7 +237,7 @@ TEST(Matrix, MulVec3Ident) {
 //    | 3 | 0 | 5 | 5 |
 //    | 0 | 8 | 3 | 8 |
 TEST(Matrix, Transpose) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		0,9,3,0,
 		9,8,0,8,
 		1,8,5,3,
@@ -245,7 +245,7 @@ TEST(Matrix, Transpose) {
 	});
 	Matrix4 result = a.transpose();
 
-	Matrix4 expect(std::array<float, 16>{
+	Matrix4 expect(std::array<double, 16>{
 		0,9,1,0,
 		9,8,8,0,
 		3,0,5,5,
@@ -268,7 +268,7 @@ TEST(Matrix, TransposeIdent) {
 //    | -3 | 2 |
 //  Then determinant(A) = 17
 TEST(Matrix, det2) {
-	Matrix2 a(std::array<float, 4>{ 1, 5, -3, 2 });
+	Matrix2 a(std::array<double, 4>{ 1, 5, -3, 2 });
 	EXPECT_EQ(a.det(), 17);
 }
 
@@ -281,14 +281,14 @@ TEST(Matrix, det2) {
 //    | -3 | 2 |
 //    |  0 | 6 |
 TEST(Matrix, Submatrix3) {
-	Matrix3 a(std::array<float, 9>{
+	Matrix3 a(std::array<double, 9>{
 		1,5,0,
 		-3,2,7,
 		0,6,-3
 	});
 	Matrix2 result = a.submatrix(0, 2);
 
-	Matrix2 expect(std::array<float, 4>{
+	Matrix2 expect(std::array<double, 4>{
 		-3,2,
 		0,6
 	});
@@ -306,7 +306,7 @@ TEST(Matrix, Submatrix3) {
 //    | -8 |  8 | 6 |
 //    | -7 | -1 | 1 |
 TEST(Matrix, Submatrix4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 	-6,1,1,6,
 	-8,5,8,6,
 	-1,0,8,2,
@@ -314,7 +314,7 @@ TEST(Matrix, Submatrix4) {
 	});
 	Matrix3 result = a.submatrix(2, 1);
 
-	Matrix3 expect(std::array<float, 9>{
+	Matrix3 expect(std::array<double, 9>{
 		-6, 1, 6,
 		-8, 8, 6,
 		-7, -1, 1
@@ -331,7 +331,7 @@ TEST(Matrix, Submatrix4) {
 //  Then determinant(B) = 25
 //    And minor(A, 1, 0) = 25
 TEST(Matrix, Minor3) {
-	Matrix3 a(std::array<float, 9>{
+	Matrix3 a(std::array<double, 9>{
 		3,5,0,
 		2,-1,-7,
 		6,-1,5
@@ -351,7 +351,7 @@ TEST(Matrix, Minor3) {
 //    And minor(A, 1, 0) = 25
 //    And cofactor(A, 1, 0) = -25
 TEST(Matrix, Cofactor3) {
-	Matrix3 a(std::array<float, 9>{
+	Matrix3 a(std::array<double, 9>{
 		3, 5, 0,
 			2, -1, -7,
 			6, -1, 5
@@ -372,7 +372,7 @@ TEST(Matrix, Cofactor3) {
 //    And cofactor(A, 0, 2) = -46
 //    And determinant(A) = -196
 TEST(Matrix, Determinant3) {
-	Matrix3 a(std::array<float, 9>{
+	Matrix3 a(std::array<double, 9>{
 		1,2,6,
 		-5,8,-4,
 		2,6,4
@@ -395,7 +395,7 @@ TEST(Matrix, Determinant3) {
 //    And cofactor(A, 0, 3) = 51
 //    And determinant(A) = -4071
 TEST(Matrix, Determinant4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		-2,-8,3,5,
 		-3,1,7,3,
 		1,2,-9,6,
@@ -417,7 +417,7 @@ TEST(Matrix, Determinant4) {
 //  Then determinant(A) = -2120
 //    And A is invertible
 TEST(Matrix, Invertible4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		6,4,4,4,
 		5,5,7,6,
 		4,-9,3,-7,
@@ -436,7 +436,7 @@ TEST(Matrix, Invertible4) {
 //  Then determinant(A) = 0
 //    And A is not invertible
 TEST(Matrix, NotInvertible4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		-4,2,-2,-3,
 		9,6,2,6,
 		0,-5,1,-5,
@@ -464,7 +464,7 @@ TEST(Matrix, NotInvertible4) {
 //      | -0.07895 | -0.22368 | -0.05263 |  0.19737 |
 //      | -0.52256 | -0.81391 | -0.30075 |  0.30639 |
 TEST(Matrix, Inverse4) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		-5,2,6,-8,
 		1,-5,1,8,
 		7,7,-6,-7,
@@ -473,11 +473,11 @@ TEST(Matrix, Inverse4) {
 	Matrix4 b = a.inverse();
 	EXPECT_EQ(a.det(), 532);
 	EXPECT_EQ(a.cofactor(2, 3), -160);
-	std::array<float, 16> array_b = b.to_array();
+	std::array<double, 16> array_b = b.to_array();
 	EXPECT_NEAR(array_b[14], -160.0 / 532.0, RAY_EPSILON);
 	EXPECT_EQ(a.cofactor(3, 2), 105);
 	EXPECT_NEAR(array_b[11], 105.0 / 532.0, RAY_EPSILON);
-	Matrix4 expect(std::array<float, 16>{
+	Matrix4 expect(std::array<double, 16>{
 		0.21805f, 0.45113f, 0.24060f, -0.04511f,
 		-0.80827f , -1.45677f , -0.44361f ,  0.52068f,
 		-0.07895f , -0.22368f , -0.05263f ,  0.19737f,
@@ -499,14 +499,14 @@ TEST(Matrix, Inverse4) {
 //    |  0.35897 |  0.35897 |  0.43590 |  0.92308 |
 //    | -0.69231 | -0.69231 | -0.76923 | -1.92308 |
 TEST(Matrix, Inverse4b) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		8, -5, 9, 2,
 		7,5,6,1,
 		-6,0,9,6,
 		-3,0,-9,-4
 	});
 	Matrix4 b = a.inverse();
-	Matrix4 expect(std::array<float, 16>{
+	Matrix4 expect(std::array<double, 16>{
 		-0.15385f, -0.15385f, -0.28205f, -0.53846f,
 			-0.07692f, 0.12308f, 0.02564f, 0.03077f,
 			0.35897f, 0.35897f, 0.43590f, 0.92308f,
@@ -528,14 +528,14 @@ TEST(Matrix, Inverse4b) {
 //    | -0.02901 | -0.14630 | -0.10926 |  0.12963 |
 //    |  0.17778 |  0.06667 | -0.26667 |  0.33333 |
 TEST(Matrix, Inverse4c) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		9, 3, 0, 9,
 			-5, -2, -6, -3,
 			-4, 9, 6, 4,
 			-7, 6, 6, 2
 	});
 	Matrix4 b = a.inverse();
-	Matrix4 expect(std::array<float, 16>{
+	Matrix4 expect(std::array<double, 16>{
 		-0.04074f, -0.07778f, 0.14444f, -0.22222f,
 			-0.07778f, 0.03333f, 0.36667f, -0.33333f,
 			-0.02901f, -0.14630f, -0.10926f, 0.12963f,
@@ -559,13 +559,13 @@ TEST(Matrix, Inverse4c) {
 //    And C ← A * B
 //  Then C * inverse(B) = A
 TEST(Matrix, InverseRoundTrip) {
-	Matrix4 a(std::array<float, 16>{
+	Matrix4 a(std::array<double, 16>{
 		3, -9, 7, 3,
 			3, -8, 2, -9,
 			-4, 4, 4, 1,
 			-6, 5, -1, 1
 	});
-	Matrix4 b(std::array<float, 16>{
+	Matrix4 b(std::array<double, 16>{
 		8, 2, 2, 2,
 			3, -1, 7, 0,
 			7, 0, 5, 4,
